daemon: Reallocate FQDN Identities after grace period
On startup the agent restores CIDR identities from the ipcache map and
allocates references to those Security Identities temporarily, until the
identity-restore-grace-period is complete. At the end of this period,
the CIDR Identities may be released. Although many of these CIDR
identities will be referenced from the FQDN selectors associated with
the DNS Name Manager, none of the references will be held by the name
manager or its FQDN selectors due to the way that internal reference
counting works for Identities in such rules. See commit de10e82639cc
("Fix identity leak via FQDN selectors") for more history on the
structure regarding identity reference counts in the tree today.

The unfortunate consequence of this structure is that once the
identity-restore-grace-period completes, these identities will be
forgotten about until the next time a DNS lookup occurs for the names
associated with these IP addresses. This can cause connectivity
disruption for applications following an upgrade or restart of the
agent, if that connectivity is enabled by ToFQDNs policy.

One might be tempted to think, well what if we just don't release those
CIDR Identity references? If we don't release them, then Cilium won't
forget how to properly enforce policy for those Identities, and traffic
will continue to flow. However, in this scenario, regardless of whether
applications continue to leverage the CIDR identities, Cilium will keep
these Identities alive persistently for the lifetime of the current
agent, and even persist them on to the next time the agent starts up.
These identities are generated based on DNS lookup activity during the
agent lifetime, and there is currently no limit on how many of these
Identities there may be. If these are not appropriately garbage
collected, then endpoint policymaps may fill more easily, leading to
incorrect policy enforcement or even rejecting the scheduling of new
endpoints.

Well, then why don't we identify exactly which IP addresses are being
used by the FQDN selectors and allocate additional references to the
corresponding CIDR Identities on startup? The current code structure
doesn't have any sort of bootstrap phase where these restored CIDRs are
identified with the context of the corresponding FQDNs. Additionally,
any particular restored Identity for an IP could be selected by multiple
selectors, and the name manager expects to retain a reference to the
identity for each selector. So it's not enough just to reference count
the identities once per CIDR. Furthermore, this means that the policy
engine must have done a full policy calculation first in order to
initialize the FQDN selectors and attach them to the name manager, so
that we can determine how many references there may be to each CIDR
Identity!

At some point we should probably take a longer harder look at this and
maybe revisit some of the design decisions from commit de10e82639cc
("Fix identity leak via FQDN selectors") around reference counting. In
the interests of preparing a safe, small, backportable fix, this commit
explores alternate short term options.

With that in mind, this commit proposes the following solution: wait
until the identity-restore-grace-period is completed and the identities
have their references released, then attempt to re-allocate the identity
references used by the name manager at that time. Any identities that
are already referenced will have referenced allocated and then removed
again by this new logic, resulting in a no-op. Any newly allocated
identities will be reinserted into the ipcache. This may cause brief
connectivity disruption to workloads impacted by the latter category of
identities, but this is better than having persistent disruption.

Signed-off-by: Joe Stringer <joe@cilium.io>