cilium: Ensure xfrm state is initialized for route IP before publish
When rolling cilium-agent or doing an upgrade while running stress test
with encryption a small number of NoStateIn errors are seen. To capture
the error state (a cilium_host IP without an xfrm state rule) you need
to get into the pod near pod init and get somewhat lucky that init
took some longer time. For example I ran `ip x s` in a pod about
15seconds after launch and captured a case with new XfrmInNoErrors,
a cilium_host ip assigned, but no xfrm state rule for it.

The conclusion is remote nodes learn the new router IP before we have
the xfrm state rule loaded. The remote nodes then start using that
IP for the IPSec tunnel outer IP resulting in the errors when they
reach the local node without the xfrm rule yet. The errors resolve
relatively quickly after init, but user is watching for errors
and trying to get lossless upgrades and restarts.

Its tempting to try and refactor NewDaemon init() to ensure we avoid
pushing any updates into the k8s node object until all the datapath
is configured. But, the code is a bit fragile and its not clear we
have enough testing to catch if this broke something. So this patch
does the minimal amount of change to get the IPsec init earlier
in the flow. This will be backported so having a minimal change
useful.

To reproduce the errors I can consistently reproduce with about
30 nodes, with httpperf test running from a pod in all nodes, and
then doing a 'rollout' of the cilium agent for awhile. Seems
a 2-3 hours almost ensures errors pop up. Usually the errors
happen much sooner.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>