cilium: Reinitialize datapath before annotating node object
As soon as we annotate the node object remote agents might pick up those
annotations and send us traffic based on the update. We need to initialize
our datapath to be prepared for that so that by the time data is received
using new information we have already prepared the datapath. Currently,
we are racing with the remote update to configure our datapath so any
delays we incur can cause us to lose and drop packets.

Interesting side effect is to know if all subsystems can have two configs
in place at the same time. The old config which might continue until the
remote node finds the update and the new one that is going to be sent
so we need to be prepared. IPSec is OK to manage the two states which
appears to be requirement to not drop traffic on update while multiple
IPs are in the network.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>